tutti i mappers funzionano

+++ src/main/java/Classes/Dessert/Dessert.java
@@ -1,17 +1,17 @@
 package Classes.Dessert;
 
 public class Dessert {
-    private final long dessertId;
+    private final int dessertId;
     private final String dessertName;
     private final double dessertPrice;
 
-    public Dessert(long id, String name, double price) {
+    public Dessert(int id, String name, double price) {
         this.dessertId = id;
         this.dessertName = name;
         this.dessertPrice = price;
     }
 
-    public long getId() {
+    public int getId() {
         return dessertId;
     }
