+++ src/main/java/Classes/Drink/Drink.java
@@ -1,17 +1,17 @@
 package Classes.Drink;
 
 public class Drink {
-    private final long drinkId;
+    private final int drinkId;
     private final String drinkName;
     private final double drinkPrice;
 
-    public Drink(long id, String name, double price) {
+    public Drink(int id, String name, double price) {
         this.drinkId = id;
         this.drinkName = name;
         this.drinkPrice = price;
     }
 
-    public long getId() {
+    public int getId() {
         return this.drinkId;
     }
